internal: Migrate entire codebase to new config crate. (#865)
* new: Create new configuration crate. (#856)

* Start on crate.

* Add toolchain configs.

* Add template.

* Pull in new crate.

* Add workspace.

* Start on project.

* Start on tasks.

* Use a trait.

* Add rel path.

* Fix lints.

* Polish.

* Polish.

* Update schematic.

* Support proto.

* Add inherited.

* Polish.

* Update again.

* Update deps.

* Add template tests.

* Add frontmatter tests.

* Start on toolchain tests.

* Update schematic.

* Update vcs.

* Update workspace.

* Update utils.

* Update cache.

* Update generator.

* Update launchpad.

* Update runner.

* Update task.

* Update project.

* Update platforms.

* Update project graph.

* Update dep graph.

* Update CLI.

* Polish.

* Update schematic.

* More updates.

* Update test utils.

* Update core.

* Update li.

* Delete old code.

* Fix lints.

* Fix tests.

* Fix cast.

* tests: Clean up all tests after config changes. (#866)

* Update vcs.

* Update workspace.

* Update generator.

* Update task.

* Update project.

* Update CLI.

* First pass.

* Second pass.

* Update task.

* Update project graph.

* Update task partials.

* Fix rebase issues.

* Fix tests.

* Update schematic.

* Path fixes.

* Fix glob on windows.

+++ crates/cli/src/commands/generate.rs
@@ -2,7 +2,7 @@ use console::Term;
 use dialoguer::{theme::Theme, Confirm, Input, MultiSelect, Select};
 use miette::IntoDiagnostic;
 use moon::load_workspace;
-use moon_config::{TemplateVariable, TemplateVariableEnumValue};
+use moon_config2::{TemplateVariable, TemplateVariableEnumValue};
 use moon_error::MoonError;
 use moon_generator::{FileState, Generator, GeneratorError, Template, TemplateContext};
 use moon_logger::{debug, map_list, trace, warn};
@@ -249,7 +249,7 @@ fn gather_variables(
                             GeneratorError::FailedToParseArgVar(name.to_owned(), e.to_string())
                         })
                         .into_diagnostic()?,
-                    None => var.default,
+                    None => var.default as i32,
                 };
 
                 if options.defaults || var.prompt.is_none() {
